Extract run_turn.rs from turn.rs (parent #680)
Fourteenth vertical-slice extraction: move the per-turn driver out of
turn.rs into a new src/agent/loop_run/run_turn.rs sibling module.

Migrated (1 pub(super) free fn):
- run_turn(&mut Agent, &str, bool, &mut InterruptMonitor) -> LoopResult
  — turn-level driver invoked from handle_user_message after the per-
  turn state resets. Pushes the user message, runs work-mode classify
  second-pass + history compaction (Act mode only), refreshes the plan
  stage, classifies action expectation, fires the pre-turn photon
  context-pack hook, and enters actor_loop_flow::run_actor_loop.

Visibility promotion in turn.rs (impl Agent method → pub(super)):
- push_user_message

No facade re-export (DR3-001). Caller update:
- handle_user_message.rs: `agent.run_turn(input, stream_output, &mut monitor)`
  → `super::run_turn::run_turn(agent, input, stream_output, &mut monitor)`

`cargo fix --tests` pruned 3 newly-unused imports from turn.rs's prelude.

turn.rs: 3,875 -> 3,817 (-58 LOC)
Cumulative from 39,489: -35,672 LOC (-90.3%)

Verification:
- cargo fmt --check
- cargo clippy --lib --all-targets -- -D warnings
- cargo test --lib (3,114 passed)
